Add file logging to HTTP/2 "unit" tests
