+++ test/Kestrel.Core.Tests/Http2StreamTests.cs
@@ -2,11 +2,8 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Buffers;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
-using System.IO.Pipelines;
 using System.Linq;
 using System.Runtime.ExceptionServices;
 using System.Threading;
@@ -15,11 +12,7 @@ using Microsoft.AspNetCore.Connections;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Features;
-using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2;
-using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http2.HPack;
-using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Infrastructure;
-using Microsoft.AspNetCore.Server.Kestrel.Transport.Abstractions.Internal;
 using Microsoft.AspNetCore.Testing;
 using Microsoft.Extensions.Logging;
 using Microsoft.Net.Http.Headers;
@@ -27,195 +20,8 @@ using Xunit;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
 {
-    public class Http2StreamTests : IDisposable, IHttpHeadersHandler
+    public class Http2StreamTests : Http2TestBase
     {
-        private static readonly string _largeHeaderValue = new string('a', HPackDecoder.MaxStringOctets);
-
-        private static readonly IEnumerable<KeyValuePair<string, string>> _browserRequestHeaders = new[]
-        {
-            new KeyValuePair<string, string>(HeaderNames.Method, "GET"),
-            new KeyValuePair<string, string>(HeaderNames.Path, "/"),
-            new KeyValuePair<string, string>(HeaderNames.Scheme, "http"),
-            new KeyValuePair<string, string>(HeaderNames.Authority, "localhost:80"),
-            new KeyValuePair<string, string>("user-agent", "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:54.0) Gecko/20100101 Firefox/54.0"),
-            new KeyValuePair<string, string>("accept", "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8"),
-            new KeyValuePair<string, string>("accept-language", "en-US,en;q=0.5"),
-            new KeyValuePair<string, string>("accept-encoding", "gzip, deflate, br"),
-            new KeyValuePair<string, string>("upgrade-insecure-requests", "1"),
-        };
-
-        private MemoryPool<byte> _memoryPool = KestrelMemoryPool.Create();
-        private DuplexPipe.DuplexPipePair _pair;
-        private readonly TestApplicationErrorLogger _logger;
-        private Http2ConnectionContext _connectionContext;
-        private Http2Connection _connection;
-        private readonly Http2PeerSettings _clientSettings = new Http2PeerSettings();
-        private readonly HPackEncoder _hpackEncoder = new HPackEncoder();
-        private readonly HPackDecoder _hpackDecoder;
-
-        private readonly ConcurrentDictionary<int, TaskCompletionSource<object>> _runningStreams = new ConcurrentDictionary<int, TaskCompletionSource<object>>();
-        private readonly Dictionary<string, string> _decodedHeaders = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-        private readonly HashSet<int> _abortedStreamIds = new HashSet<int>();
-        private readonly object _abortedStreamIdsLock = new object();
-
-        private readonly RequestDelegate _noopApplication;
-        private readonly RequestDelegate _echoMethod;
-        private readonly RequestDelegate _echoHost;
-        private readonly RequestDelegate _echoPath;
-        private readonly RequestDelegate _waitForAbortApplication;
-        private readonly RequestDelegate _waitForAbortFlushingApplication;
-        private readonly RequestDelegate _waitForAbortWithDataApplication;
-
-        private Task _connectionTask;
-
-        public Http2StreamTests()
-        {
-            _noopApplication = context => Task.CompletedTask;
-
-            _echoMethod = context =>
-            {
-                context.Response.Headers["Method"] = context.Request.Method;
-
-                return Task.CompletedTask;
-            };
-
-            _echoHost = context =>
-            {
-                context.Response.Headers[HeaderNames.Host] = context.Request.Headers[HeaderNames.Host];
-
-                return Task.CompletedTask;
-            };
-
-            _echoPath = context =>
-            {
-                context.Response.Headers["path"] = context.Request.Path.ToString();
-                context.Response.Headers["rawtarget"] = context.Features.Get<IHttpRequestFeature>().RawTarget;
-
-                return Task.CompletedTask;
-            };
-
-            _waitForAbortApplication = async context =>
-            {
-                var streamIdFeature = context.Features.Get<IHttp2StreamIdFeature>();
-                var sem = new SemaphoreSlim(0);
-
-                context.RequestAborted.Register(() =>
-                {
-                    lock (_abortedStreamIdsLock)
-                    {
-                        _abortedStreamIds.Add(streamIdFeature.StreamId);
-                    }
-
-                    sem.Release();
-                });
-
-                await sem.WaitAsync().DefaultTimeout();
-
-                _runningStreams[streamIdFeature.StreamId].TrySetResult(null);
-            };
-
-            _waitForAbortFlushingApplication = async context =>
-            {
-                var streamIdFeature = context.Features.Get<IHttp2StreamIdFeature>();
-                var sem = new SemaphoreSlim(0);
-
-                context.RequestAborted.Register(() =>
-                {
-                    lock (_abortedStreamIdsLock)
-                    {
-                        _abortedStreamIds.Add(streamIdFeature.StreamId);
-                    }
-
-                    sem.Release();
-                });
-
-                await sem.WaitAsync().DefaultTimeout();
-
-                await context.Response.Body.FlushAsync();
-
-                _runningStreams[streamIdFeature.StreamId].TrySetResult(null);
-            };
-
-            _waitForAbortWithDataApplication = async context =>
-            {
-                var streamIdFeature = context.Features.Get<IHttp2StreamIdFeature>();
-                var sem = new SemaphoreSlim(0);
-
-                context.RequestAborted.Register(() =>
-                {
-                    lock (_abortedStreamIdsLock)
-                    {
-                        _abortedStreamIds.Add(streamIdFeature.StreamId);
-                    }
-
-                    sem.Release();
-                });
-
-                await sem.WaitAsync().DefaultTimeout();
-
-                await context.Response.Body.WriteAsync(new byte[10], 0, 10);
-
-                _runningStreams[streamIdFeature.StreamId].TrySetResult(null);
-            };
-
-            _hpackDecoder = new HPackDecoder((int)_clientSettings.HeaderTableSize);
-
-            _logger = new TestApplicationErrorLogger();
-
-            InitializeConnectionFields(KestrelMemoryPool.Create());
-        }
-
-        private void InitializeConnectionFields(MemoryPool<byte> memoryPool)
-        {
-            _memoryPool = memoryPool;
-
-            // Always dispatch test code back to the ThreadPool. This prevents deadlocks caused by continuing
-            // Http2Connection.ProcessRequestsAsync() loop with writer locks acquired. Run product code inline to make
-            // it easier to verify request frames are processed correctly immediately after sending the them.
-            var inputPipeOptions = new PipeOptions(
-                pool: _memoryPool,
-                readerScheduler: PipeScheduler.Inline,
-                writerScheduler: PipeScheduler.ThreadPool,
-                useSynchronizationContext: false
-            );
-            var outputPipeOptions = new PipeOptions(
-                pool: _memoryPool,
-                readerScheduler: PipeScheduler.ThreadPool,
-                writerScheduler: PipeScheduler.Inline,
-                useSynchronizationContext: false
-            );
-
-            _pair = DuplexPipe.CreateConnectionPair(inputPipeOptions, outputPipeOptions);
-
-            _connectionContext = new Http2ConnectionContext
-            {
-                ConnectionFeatures = new FeatureCollection(),
-                ServiceContext = new TestServiceContext()
-                {
-                    Log = new TestKestrelTrace(_logger),
-                },
-                MemoryPool = _memoryPool,
-                Application = _pair.Application,
-                Transport = _pair.Transport
-            };
-
-            _connection = new Http2Connection(_connectionContext);
-        }
-
-        public void Dispose()
-        {
-            _pair.Application.Input.Complete();
-            _pair.Application.Output.Complete();
-            _pair.Transport.Input.Complete();
-            _pair.Transport.Output.Complete();
-            _memoryPool.Dispose();
-        }
-
-        void IHttpHeadersHandler.OnHeader(Span<byte> name, Span<byte> value)
-        {
-            _decodedHeaders[name.GetAsciiStringNonNullCharacters()] = value.GetAsciiStringNonNullCharacters();
-        }
-
         [Fact]
         public async Task HEADERS_Received_EmptyMethod_Reset()
         {
@@ -1105,7 +911,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
                 withFlags: (byte)Http2DataFrameFlags.END_STREAM,
                 withStreamId: 1);
 
-            Assert.Contains(_logger.Messages, m => m.Exception?.Message.Contains("Response Content-Length mismatch: too many bytes written (12 of 11).") ?? false);
+            Assert.Contains(TestApplicationErrorLogger.Messages, m => m.Exception?.Message.Contains("Response Content-Length mismatch: too many bytes written (12 of 11).") ?? false);
 
             await StopConnectionAsync(expectedLastStreamId: 1, ignoreNonGoAwayFrames: false);
 
@@ -1182,7 +988,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
                 withFlags: (byte)Http2DataFrameFlags.END_STREAM,
                 withStreamId: 1);
 
-            Assert.Contains(_logger.Messages, m => m.Exception?.Message.Contains("Response Content-Length mismatch: too few bytes written (0 of 11).") ?? false);
+            Assert.Contains(TestApplicationErrorLogger.Messages, m => m.Exception?.Message.Contains("Response Content-Length mismatch: too few bytes written (0 of 11).") ?? false);
 
             await StopConnectionAsync(expectedLastStreamId: 1, ignoreNonGoAwayFrames: false);
 
@@ -1538,7 +1344,7 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
                 withFlags: (byte)Http2DataFrameFlags.END_STREAM,
                 withStreamId: 1);
 
-            Assert.Contains(_logger.Messages, m => (m.Exception?.Message.Contains("App Faulted") ?? false) && m.LogLevel == LogLevel.Error);
+            Assert.Contains(TestApplicationErrorLogger.Messages, m => (m.Exception?.Message.Contains("App Faulted") ?? false) && m.LogLevel == LogLevel.Error);
 
             await StopConnectionAsync(expectedLastStreamId: 1, ignoreNonGoAwayFrames: false);
 
@@ -1858,221 +1664,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Tests
             Assert.Equal("The request was aborted", thrownEx.Message);
             Assert.IsType<ConnectionAbortedException>(thrownEx.InnerException);
             Assert.Equal(CoreStrings.ConnectionAbortedByApplication, thrownEx.InnerException.Message);
-        }
-
-        private async Task InitializeConnectionAsync(RequestDelegate application)
-        {
-            _connectionTask = _connection.ProcessRequestsAsync(new DummyApplication(application));
-
-            await SendPreambleAsync().ConfigureAwait(false);
-            await SendSettingsAsync();
-
-            await ExpectAsync(Http2FrameType.SETTINGS,
-                withLength: 6,
-                withFlags: 0,
-                withStreamId: 0);
-
-            await ExpectAsync(Http2FrameType.SETTINGS,
-                withLength: 0,
-                withFlags: (byte)Http2SettingsFrameFlags.ACK,
-                withStreamId: 0);
-        }
-
-        private async Task StartStreamAsync(int streamId, IEnumerable<KeyValuePair<string, string>> headers, bool endStream)
-        {
-            var tcs = new TaskCompletionSource<object>(TaskCreationOptions.RunContinuationsAsynchronously);
-            _runningStreams[streamId] = tcs;
-
-            var frame = new Http2Frame();
-            frame.PrepareHeaders(Http2HeadersFrameFlags.NONE, streamId);
-            var done = _hpackEncoder.BeginEncode(headers, frame.HeadersPayload, out var length);
-            frame.Length = length;
-
-            if (done)
-            {
-                frame.HeadersFlags = Http2HeadersFrameFlags.END_HEADERS;
-            }
-
-            if (endStream)
-            {
-                frame.HeadersFlags |= Http2HeadersFrameFlags.END_STREAM;
-            }
-
-            await SendAsync(frame.Raw);
-
-            while (!done)
-            {
-                frame.PrepareContinuation(Http2ContinuationFrameFlags.NONE, streamId);
-                done = _hpackEncoder.Encode(frame.HeadersPayload, out length);
-                frame.Length = length;
-
-                if (done)
-                {
-                    frame.ContinuationFlags = Http2ContinuationFrameFlags.END_HEADERS;
-                }
-
-                await SendAsync(frame.Raw);
-            }
-        }
-
-        private Task WaitForAllStreamsAsync()
-        {
-            return Task.WhenAll(_runningStreams.Values.Select(tcs => tcs.Task)).DefaultTimeout();
-        }
-
-        private Task SendAsync(ReadOnlySpan<byte> span)
-        {
-            var writableBuffer = _pair.Application.Output;
-            writableBuffer.Write(span);
-            return FlushAsync(writableBuffer);
-        }
-
-        private static async Task FlushAsync(PipeWriter writableBuffer)
-        {
-            await writableBuffer.FlushAsync();
-        }
-
-        private Task SendPreambleAsync() => SendAsync(new ArraySegment<byte>(Http2Connection.ClientPreface));
-
-        private Task SendSettingsAsync()
-        {
-            var frame = new Http2Frame();
-            frame.PrepareSettings(Http2SettingsFrameFlags.NONE, _clientSettings.GetNonProtocolDefaults());
-            return SendAsync(frame.Raw);
-        }
-
-        private async Task<bool> SendHeadersAsync(int streamId, Http2HeadersFrameFlags flags, IEnumerable<KeyValuePair<string, string>> headers)
-        {
-            var frame = new Http2Frame();
-
-            frame.PrepareHeaders(flags, streamId);
-            var done = _hpackEncoder.BeginEncode(headers, frame.Payload, out var length);
-            frame.Length = length;
-
-            await SendAsync(frame.Raw);
-
-            return done;
-        }
-
-        private Task SendDataAsync(int streamId, Span<byte> data, bool endStream)
-        {
-            var frame = new Http2Frame();
-
-            frame.PrepareData(streamId);
-            frame.Length = data.Length;
-            frame.DataFlags = endStream ? Http2DataFrameFlags.END_STREAM : Http2DataFrameFlags.NONE;
-            data.CopyTo(frame.DataPayload);
-
-            return SendAsync(frame.Raw);
-        }
-
-        private Task SendRstStreamAsync(int streamId)
-        {
-            var rstStreamFrame = new Http2Frame();
-            rstStreamFrame.PrepareRstStream(streamId, Http2ErrorCode.CANCEL);
-            return SendAsync(rstStreamFrame.Raw);
-        }
-
-        private async Task<Http2Frame> ReceiveFrameAsync()
-        {
-            var frame = new Http2Frame();
-
-            while (true)
-            {
-                var result = await _pair.Application.Input.ReadAsync();
-                var buffer = result.Buffer;
-                var consumed = buffer.Start;
-                var examined = buffer.End;
-
-                try
-                {
-                    Assert.True(buffer.Length > 0);
-
-                    if (Http2FrameReader.ReadFrame(buffer, frame, 16_384, out consumed, out examined))
-                    {
-                        return frame;
-                    }
-
-                    if (result.IsCompleted)
-                    {
-                        throw new IOException("The reader completed without returning a frame.");
-                    }
-                }
-                finally
-                {
-                    _pair.Application.Input.AdvanceTo(consumed, examined);
-                }
-            }
-        }
-
-        private async Task<Http2Frame> ExpectAsync(Http2FrameType type, int withLength, byte withFlags, int withStreamId)
-        {
-            var frame = await ReceiveFrameAsync().DefaultTimeout();
-
-            Assert.Equal(type, frame.Type);
-            Assert.Equal(withLength, frame.Length);
-            Assert.Equal(withFlags, frame.Flags);
-            Assert.Equal(withStreamId, frame.StreamId);
-
-            return frame;
-        }
-
-        private Task StopConnectionAsync(int expectedLastStreamId, bool ignoreNonGoAwayFrames)
-        {
-            _pair.Application.Output.Complete();
-
-            return WaitForConnectionStopAsync(expectedLastStreamId, ignoreNonGoAwayFrames);
-        }
-
-        private Task WaitForConnectionStopAsync(int expectedLastStreamId, bool ignoreNonGoAwayFrames)
-        {
-            return WaitForConnectionErrorAsync<Exception>(ignoreNonGoAwayFrames, expectedLastStreamId, Http2ErrorCode.NO_ERROR, expectedErrorMessage: null);
-        }
-
-        private async Task WaitForConnectionErrorAsync<TException>(bool ignoreNonGoAwayFrames, int expectedLastStreamId, Http2ErrorCode expectedErrorCode, string expectedErrorMessage)
-            where TException : Exception
-        {
-            var frame = await ReceiveFrameAsync();
-
-            if (ignoreNonGoAwayFrames)
-            {
-                while (frame.Type != Http2FrameType.GOAWAY)
-                {
-                    frame = await ReceiveFrameAsync();
-                }
-            }
-
-            Assert.Equal(Http2FrameType.GOAWAY, frame.Type);
-            Assert.Equal(8, frame.Length);
-            Assert.Equal(0, frame.Flags);
-            Assert.Equal(0, frame.StreamId);
-            Assert.Equal(expectedLastStreamId, frame.GoAwayLastStreamId);
-            Assert.Equal(expectedErrorCode, frame.GoAwayErrorCode);
-
-            if (expectedErrorMessage != null)
-            {
-                var message = Assert.Single(_logger.Messages, m => m.Exception is TException);
-                Assert.Contains(expectedErrorMessage, message.Exception.Message);
-            }
-
-            await _connectionTask;
-            _pair.Application.Output.Complete();
-        }
-
-        private async Task WaitForStreamErrorAsync(int expectedStreamId, Http2ErrorCode expectedErrorCode, string expectedErrorMessage)
-        {
-            var frame = await ReceiveFrameAsync();
-
-            Assert.Equal(Http2FrameType.RST_STREAM, frame.Type);
-            Assert.Equal(4, frame.Length);
-            Assert.Equal(0, frame.Flags);
-            Assert.Equal(expectedStreamId, frame.StreamId);
-            Assert.Equal(expectedErrorCode, frame.RstStreamErrorCode);
-
-            if (expectedErrorMessage != null)
-            {
-                Assert.Contains(_logger.Messages, m => m.Exception?.Message.Contains(expectedErrorMessage) ?? false);
-            }
         }
     }
 }
