+++ test/Kestrel.InMemory.FunctionalTests/Kestrel.InMemory.FunctionalTests.csproj
@@ -9,7 +9,6 @@
 
   <ItemGroup>
     <Compile Include="..\shared\*.cs" LinkBase="shared" />
-    <Compile Include="..\shared\FunctionalTestHelpers\*.cs" LinkBase="shared\FunctionalTestHelpers" />
     <Content Include="..\shared\TestCertificates\*.pfx" LinkBase="shared\TestCertificates" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
