+++ test/Kestrel.Transport.Libuv.BindTests/Kestrel.Transport.Libuv.BindTests.csproj
@@ -12,7 +12,6 @@
     <Compile Include="..\Kestrel.Transport.BindTests\**\*.cs" />
     <Compile Include="..\Kestrel.Transport.Libuv.FunctionalTests\TransportSelector.cs" />
     <Compile Include="..\shared\*.cs" LinkBase="shared" />
-    <Compile Include="..\shared\FunctionalTestHelpers\*.cs" LinkBase="shared\FunctionalTestHelpers" />
     <Compile Include="..\shared\TransportTestHelpers\*.cs" LinkBase="shared\TransportTestHelpers" />
     <Content Include="..\shared\TestCertificates\*.pfx" LinkBase="shared\TestCertificates" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
