+++ test/shared/DiagnosticMemoryPoolFactory.cs
